requirements.txt
requests
pyyaml
pandas
matplotlib
mplfinance
python-docx
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
tenacity
yfinance
certifi
